fix lazymatrix implementations, all tests are passing.

+++ src/utils/math/blas/Constants.h
@@ -11,7 +11,7 @@
 
 namespace MyBLAS {
 
-using NumericType = float;
+using NumericType = long double;
 
 }
